// aah_apr_front_end-master/types/Types.ts
export type SelectedCommunityDialoguesGroups = {
  communityDialogueId: string;
  group: {
    name: string;
    id: string;
  };
}[];

export type CommunityDialogues = {
  id: string;
  name: string;
  groups: {
    id: string;
    name: string;
  }[];
}[];